bugfix: decode frame; display correct image
